Optimized: Queries like `{ signatures: { any: { node_id: { in: ["1", "2"] } } } }` generates   AQL like `(doc.signatures[*].node_id IN @v1) OR (doc.signatures[*].node_id IN @v2)`.
