+++ package.json
@@ -1,6 +1,6 @@
 {
 	"name": "ton-q-server",
-	"version": "0.27.0",
+	"version": "0.27.1",
 	"description": "TON Q Server – realtime queries over TON blockchain.",
 	"main": "server/index.js",
 	"repository": "git@github.com:tonlabs/ton-q-server.git",
